% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_data.R
\name{baseline_data}
\alias{baseline_data}
\title{Baseline pupil data to the average pupil size within a window}
\usage{
baseline_data(data, pupil, start, stop)
}
\arguments{
\item{data}{a PupillometrySENDA dataframe}

\item{pupil}{a column name denoting pupil data}

\item{start}{start time of baseline window}

\item{stop}{stop time of baseline window}
}
\value{
A PupillometrySENDA dataframe, with baselined pupil
}
\description{
This function is for use with the PupillometrySENDA package to baseline each participant's pupil size to the
mean pupil size within a window.
This may not be necessary if you are doing purely within-subject analyses, but it is convenient for
comparison across subjects, and makes results more uniform.
}
\examples{
Sdata <- make_pupillometrysenda_data(data = pupil_data,
subject = "ID",
trial = "Trial",
time = "Time",
condition = "Type")
base_data <- baseline_data(data = Sdata, 
pupil = "LPupil", 
start = 0, 
stop = 100)
}
