% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_missing_data.R
\name{clean_missing_data}
\alias{clean_missing_data}
\title{Clean missing data above an acceptable threshold}
\usage{
clean_missing_data(
  data,
  pupil,
  trial_threshold = 1,
  subject_trial_threshold = 1
)
}
\arguments{
\item{data}{your data of class PupillometrySENDA}

\item{pupil}{a column name denoting pupil size}

\item{trial_threshold}{a proportion of missing data over which a trial can be considered lost}

\item{subject_trial_threshold}{a proportion of missing trials over which a participant can be considered lost.}
}
\value{
A cleaned PupillometrySENDA dataframe
}
\description{
This function can be used to remove trials and participants
who do not meet the threshold for a study. Note that there are two parameters for
cleaning, one to remove trials above a threshold,
the second to remove participants who drop more than a certain amount of trials.
}
\examples{
Sdata <- make_pupillometrysenda_data(data = pupil_data,
subject = "ID",
trial = "Trial",
time = "Time",
condition = "Type")

Sdata2 <- clean_missing_data(data = Sdata, 
pupil = "LPupil", 
trial_threshold = 0.8, 
subject_trial_threshold = 0.8)
}
