% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_pupil_analysis.R
\name{total_pupil_analysis}
\alias{total_pupil_analysis}
\title{Perform preprocessing on data frames for PupillometrySENDA}
\usage{
total_pupil_analysis(
  data,
  pupil,
  sampling_rate,
  trial_threshold,
  subject_trial_threshold,
  baseline_start,
  baseline_end
)
}
\arguments{
\item{data}{Dataframe for PupillometrySENDA}

\item{pupil}{Pupil data column name}

\item{sampling_rate}{Sampling frequency (unit: ms)}

\item{trial_threshold}{Allowable loss ratio per trial}

\item{subject_trial_threshold}{Allowable missing percentages per subject}

\item{baseline_start}{Starting point of baseline correction}

\item{baseline_end}{End point of baseline correction}
}
\value{
A dataframe for PupillometrySENDA after preprocessing
}
\description{
Downsampling, filtering, blink removal, rejection by missing values, interpolation of missing data,
and baseline correction are performed on the data frames for PupillometrySENDA.
}
\examples{
Sdata <- make_pupillometrysenda_data(data,
subject = "ID",
trial = "Trial",
time = "Time",
condition = "Type")
result_data <- total_pupil_analysis(data = Sdata,
pupil = "LPupil",
sampling_rate = 200,
trial_threshold = 0.75,
subject_trial_threshold = 0.75,
baseline_start = 800,
baseline_end = 1000)

}
